--- src/main/kotlin/Apikey.kt
@@ -1,3 +1,0 @@
-object Apikey {
-    val miToken = "1794779463:AAGbSW8ag5Y7YbppbnII4njkrBAKqXNJLaQ"
-}
